Move extensions browser into main viewport
