+++ src/app/features/code/CodeWorkspaceMainView.tsx
@@ -1,3 +1,4 @@
+import { Suspense, lazy } from "react";
 import type { DesktopAction } from "../../desktop/actions";
 import type {
   AppSettings,
@@ -9,6 +10,11 @@ import type { Project, View } from "../../types";
 import { LandingView } from "../../views/LandingView";
 import { SettingsView } from "../../views/SettingsView";
 import { ThreadView } from "../../views/ThreadView";
+
+const ExtensionsView = lazy(async () => {
+  const module = await import("../../views/ExtensionsView");
+  return { default: module.ExtensionsView };
+});
 
 type CodeWorkspaceMainViewProps = {
   activeView: View;
@@ -70,6 +76,23 @@ export function CodeWorkspaceMainView({
         projects={projects}
         onAction={onAction}
       />
+    );
+  }
+
+  if (activeView === "extensions") {
+    return (
+      <Suspense
+        fallback={
+          <div className="mx-auto grid h-full w-full max-w-[760px] content-start gap-4 px-2 pt-6 pb-6">
+            <div className="grid gap-1">
+              <h1 className="m-0 text-[18px] font-medium text-[color:var(--text)]">Extensions</h1>
+              <p className="m-0 text-[13px] text-[color:var(--muted)]">Loading packages…</p>
+            </div>
+          </div>
+        }
+      >
+        <ExtensionsView />
+      </Suspense>
     );
   }
