+++ src/app/types.ts
@@ -8,7 +8,7 @@ export type {
   TurnDiffSummary,
 } from "../../shared/desktop-contracts.js";
 
-export type View = "code" | "thread" | "chat" | "claw" | "work" | "settings";
+export type View = "code" | "thread" | "chat" | "claw" | "work" | "settings" | "extensions";
 
 export type FeatureCard = {
   title: string;
